Implement Ant Colony Optimization ( ACO )

+++ src/main/java/com/deliverif/app/algorithm/astar/Scorer.java
@@ -1,5 +1,5 @@
 package com.deliverif.app.algorithm.astar;
 
 public interface Scorer<T extends GraphNode> {
-    double computeCost(T from, T to);
+    float computeCost(T from, T to);
 }
